requirements.txt
numpy>=1.21.0
scikit-learn>=1.0.2
Pillow>=9.0.0
tkinterdnd2>=0.3.0 